Fixed all lint errors
