+++ src/App/Blog/entryPath.dto.ts
@@ -1,0 +1,25 @@
+import { Field, InputType, Int, ObjectType } from '@nestjs/graphql'
+
+@InputType()
+export class EntryPathInput {
+  @Field(() => String)
+  slug = ''
+
+  @Field(() => String)
+  path = ''
+
+  @Field(() => Int)
+  timestamp = 0
+}
+
+@ObjectType()
+export class EntryPathOutput {
+  @Field(() => String)
+  slug = ''
+
+  @Field(() => String)
+  path = ''
+
+  @Field(() => Int)
+  timestamp = 0
+}
